core/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
paho-mqtt==2.1.0
pydantic==2.8.2
